opensm: Add better error output when parsing node name maps
Presently, any error in opening a node name map file leads
to a generic:

"WARNING failed to open node name map"

error.  This patch adds better error output when the issue
is specifically parsing errors in the file.  The user
will get:

"WARNING failed to parse node name map"

and an output of the line with the particular issue.

Signed-off-by: Albert Chu <chu11@llnl.gov>
Signed-off-by: Alex Netes <alexne@mellanox.com>