OpenSM: DFSSSP does not find LIDs due to wrong byte order (v2)
Problem:
argument list for external calls of path_sl(...) haas been changed at some point in the past;
path_sl(...) arguments for slid/dlid are now in network byte order;
internal storage of lids is host byte order;
this mismatch results in a return value of 'hint_for_default_sl' of DFSSSP's get_dfsssp_sl function for every request

Fix:
lids will be stored in network byte order, so that a conversion is not necessary and
DFSSSP returns the correct SL for that request

This is version 2 of the original patch, because I forgot to change some internal calls.
Please use this patch, instead of the patch from December 17.

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Alex Netes <alexne@mellanox.com>