opensm/osm_console.c: Support portstatus output for unenabled width/speed
On occasion, it has been observed that active link width and/or speed
is set at a level not enabled.  For example, a link may be active at
QDR but SDR & DDR speeds are the only ones enabled.

Under these scenarios, the portstatus option in the console may
incorrectly output port status, which can confuse users.

A new port output status of "unenabled width" or "unenabled speed" is
output when this situation is discovered, helping users find problems
in the fabric.

Signed-off-by: Albert Chu <chu11@llnl.gov>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>