opensm/osm_torus.c: avoid the possibility of following stale ->priv pointers
Torus-2QoS makes persistent use of osm_switch_t:priv and osm_port_t:priv
to speed its routing calculations.

In some situations t2q is unable to determine the location in the torus
topology of a discovered switch.  An example of this occurs when all the
inter-switch links but one connected to a switch have failed.

When this happens, at the end of a full sweep opensm will attempt to
program the VLArb tables for that switch.  Torus_update_osm_vlarb()
will follow a stale osm_switch_t:priv pointer, left over from the
previous routing cycle when switch was discovered in the torus,
ultimately resulting in a segmentation fault.

Prevent this by clearing stale values of osm_switch_t:priv as part of
the first step in torus topology discovery.  Do the same thing for
osm_port_t:priv, even though it has not yet been implicated in this
sort of issue yet.

Signed-off-by: Jim Schutt <jaschut@sandia.gov>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>