Clean up event subscriptions if a port goes away
Event subscriptions needs to be cleaned up if a port goes away.
If the port comes online again later it may no longer want to
receive the events on the same QPN. If the "old" QPN is used for
something else the SM forwarding events may cause QKey violations.

This behavior is made configurable and it needs to be explicitly
enabled (default is off).

Signed-off-by: Line Holen <Line.Holen@oracle.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>