dfsssp: send multicast forwarding tables to switches
Issue: root switch of the mcast spanning tree was ignored.
When a port of the root switch is part of the mcast group, then
it won't be processed and non of its ports will be part of
the resulting mcast forwarding table.

Fix: remove the test for used_link==NULL, because all switches in
adj_list should have a used_link set by the prior dijkstra step
(except the root switch) => test not needed and root switch will
be included in mcast update.

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>