Better handling of topology changes in the fabric
The patch tries to solve the following problem:
When a newly discovered switch is rebooted during the configuration cycle,
SM ends up setting all Initialized ports to Active, but the configuration
on the switch such as Pkey tables, QoS, etc. might be incorrect.

The fix solves this in two steps. First, it turns on the need_update flag
when switch's StateChange bit is detected ON or CA's neighbor switch has
StateChange bit ON. Second, it clears the StateChange bit on the switches,
before any configuration is done. This assures that we don't miss changes
in the fabric. If a switch was rebooted during a sweep, we will detect it
in a subsequent sweep and configure all its' neighbors from scratch.

Signed-off-by: Alex Netes <alexne@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>