dfsssp: avoid crash due to port initialisation errors
If a subset of ports in the fabric are connected but their
initialisation failed, then DFSSSP might run into an 'assertion
failed' exception within the deadlock removal part and opensm
terminates.

Instead of iterating over p_mgr->p_subn->port_guid_tbl we iterate
now over p_mgr->port_order_list, which contains valid ports only.

We also add a 'return 1' in update_lft() for the case the previous
Dijkstra step failed to set an valid forwarding port for a switch,
which is considered an error in the routing algorithm and OpenSM
should stop DFSSSP at this state.

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>