osm_node_info_rcv.c: Fix multiple switch discovery during a sweep
Because of the race between Get(SwitchInfo) and Get(NodeInfo) to the
same switch, SM will send additional Get(SwitchInfo) before it gets
first reply. In this sequence, SM sends many redundant MADs and slows
down the discovery process significantly.

With this fix, when a switch doesn't respond to Get(SwitchInfo) once
during the sweep it will be dropped.

Signed-off-by: Alex Netes <alexne@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>