osm_sminfo_rcv.c: Send trap 144 to a newly found MASTER SM when in MASTER state
Before this patch, when SM in Master state finds other Master SM, it
sends trap 144 to previously found Master SM/SM with higher priority when
it was in Discovering/Standby state.

This can lead to wrong behaviour in a multi-SM topology:
Setup: SM1 with priority 1, SM2 with priority 2, SM3 with priority 3.
Flow:
1. setting SM3 to ignore SMInfo MADs -> SM2 become master
2. setting SM2 to ignore SMInfo MADs -> SM1 become master
3. setting SM2 to accept SMInfo MADs
4. SM2 sends SMInfo to SM1 -> finds that SM1 is master
5. SM2 sends trap 144 to SM3 instead of sending it to SM1

Signed-off-by: Alex Netes <alexne@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>