perfmgr: Add xmit_wait support
xmit_waits can be prevalent in an oversubscribed or otherwise congested fabric.
Therefore 2 options are added to the config file to control the logging of
these errors.  Errors are always collected if support is indicated by
ClassPortInfo and can be seen within the console output of the perfmgr.

perfmgr_xmit_wait_log FALSE
perfmgr_xmit_wait_threshold 65535

Signed-off-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>