osm_perfmgr.c: Remove unnecessary log message
The "PM state %d was NOT Suspended???" log message is unnecessary.
It is not an error, and in fact is quite normal for the sweep_state
to be not suspended.

Multiple threads can enter perfmgr_send_mad() and sit on the
conditional wait.  The first one to be signaled will obtain the
perfmgr lock and set the sweep_state to ACTIVE.  Every other thread
sitting on the conditional wait that is signaled afterwards has a
very high chance of seeing the sweep_state in ACTIVE instead of SUSPENDED.

Signed-off-by: Albert L. Chu <chu11@llnl.gov>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>