Change osm_subn_t.log_max_size type to uint32_t
Change log_max_size attribute type in order to avoid problems when
parsing the parameter on big endian systems.

Bug description:

opts_parse_uint32() is used to parse the log_max_size parameter which
sets the value of the variable by assigning it to the address of the
variable that should hold the parameter's value.

Since p_opt->log_max_size is unsigned long, opts_parse_uint32 assigned
the value only to the first 32 bits of the variable, which on big endian
systems are the most significant bits.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>