osm_mcast_mgr.c: Invalidate cache due to multicast routing errors
When running OpenSM with Unicast Cache option and a leaf switch is
removed from the fabric, minhop table might have entries that lead to
disconnected ports, and multicast routing might fail.

With this change, when OpenSM tries to route multicast tree through a
disconnected port, it invalidates the cache, and forces fabric
rerouting.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>