Fix turning on first_time_master_sweep flag
Race condition may occur during very long heavy sweep caused by
multiple timeouts in the fabric when Master SM is stuck in fabric configuration stage.

If Master SM gets temporarily isolated from some part of the fabric,
Standby SM may became Master SM in this part of the fabric
and reconfigure it, including sending client reregister to the
hosts belonging to this part of the fabric.

After that the following race may occur:
1. The connectivity to the lost part of the fabric is re-established but Master SM
is still busy with timeouts
2. Standby SM (now Master in this part of the fabric) sends Handover notification to the Master SM
3. Master SM receives the Handover notification and sets first_time_master_sweep to TRUE
4. This setting is overridden to FALSE at the end of the long heavy sweep of the Master SM
5. Next immediate heavy sweep is not "first time master sweep". So that PortInfo MAD including
Master SM LID is sent to the hosts of previously disconnected fabric but client_rereg bit is not set.

As a result, IPoIB does not work correctly on these hosts as SM LID is not updated in ipoib driver.
(This happens due to known issue in low-level mlnx drivers when SM_CHANGE event is not generated upon
SM LID change without client_reregister bit being set in PortInfo sent by opensm).

The fix sets first_time_master_sweep and other flags at the beginning of the
heavy sweep and avoids the race in turning on first_time_master_sweep flag.

Signed-off-by: Vladimir Koushnir <vladimirk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>