osm_pkey_mgr.c: Use calloc instead of malloc in pkey_mgr_process_physical_port
An allocated list item is not initialized (set to 0). After filling it's
value, we try to add it to a list. However, the adding method includes
an ASSERT in case that the item already points to the list it is required
to be added to. So in order to avoid a case where a place in the memory
that was used for an item from this list is reallocated and tried to be
added again, we need to make sure that the new item that we allocated is
initialized and not pointing by mistake to any list.

Signed-off-by: Shlomi Nimrodi <shlomin@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>