osm_lid_mgr.c: Send client reregistration in case of SM LID change
If, for some reason, the SM LID recorded in PortInfo is different from
our SM LID, it is safer to send client reregister to this port
to trigger reregistration of the node in Master SM.

This scenario may occur when Standby SM loses connectivity to the Master SM
and becomes Master SM for some time period. If the connectivity is
re-established but Handover sent by Standby SM is lost. the only indication
about SM handovers in the fabric is SM LID mis-configuration in some nodes.

It is important to note that this is a workaround.
Some applications/ULPs only look at client reregistration and not at
SM LID change.
Most IB device drivers are not currently issuing SM LID change local event.

This fix changes opensm to set client reregister bit in PortInfo(Set)
in the scenario above.

Signed-off-by: Vladimir Koushnir <vladimirk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>