osm_switch.c: Fix potential memory leak due to misuse of realloc
In the function osm_switch_prepare_path_rebuild, OpenSM reallocates
memory for p_sw->new_lft using realloc, and overwrites the original
pointer's value with the result without verifying that it succeeded.
Hence, if realloc failed, the memory that was previously allocated for
p_sw->new_lft can not be freed.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>