osm_perfmgr.c: Reset physp_discovered before discovery
Bug description:
When PerfMgr starts discovery, it resets the discovery_count of ports and
node, but it doesn't reset the physp_discoved buffers.
As a result, the drop manager frees osm_port_t and osm_physp_t objects
of nodes without removing the nodes themselves.
On the next discovery process, when OpenSM receives responses for
NodeInfo from a switch, it treats it as a NodeInfo from an existing
node and assumes that osm_physp_t objects of the switch exists and tries
to access them.

This change adds code that clears the physp_discoverd buffers of all nodes
at the beginning of the discovery process in PerfMgr.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>