osm_ucast_mgr.c: Support diverse paths for LMC > 0 when scatter_ports is enabled
The patch adds an ability to calculate diverse paths for all LIDs
(LMC > 0) derived from the same Base LID,
when scatter ports is enabled (different from 0).
We apply randomization during selection of the outports
only when routing is calculated for the Base LID. For LIDs different
from the Base LID, we don't randomize the outport but apply the procedure
for calculating diverse path as if scatter_ports equals 0.
The patch is relevant for minhop/updn routing.

Signed-off-by: Vladimir Koushnir <vladimirk@mellanox.com>
Signed-off-by: Shlomi Nimrodi <shlomin@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>