complib: Add a d-ary heap
This patch adds a d-ary heap in the component library, which
can be used by routing algorithms, e.g., as support for Dijkstra's
algorithm, and others parts of opensm (if some version of priority
queue is needed).

The heap is stored implicitly in an array and the user can
specify the branching factor 'd'. Furthermore, the default
minimum heap can be changed to a maximum heap with a user
supplied compare function.

Supported heap functions are: insert, delete, extractMin, and
decreaseKey. Additinal functions are available to read heap
attributes, such as getSize or getCapasity, or to modify the heap,
such as resize.

The memory for the d-ary heap is allocated in the init function.
The only administrative action by the user is to store the heap
indices somewhere and specify a function to be used by the heap
to update the indices when elements are swapped in the heap.
The index is necessary for the user to apply changes for elements
in the heap with decreaseKey and delete operations.

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>