dfsssp: Replace the internal heap implementation
This patch removes the internal binary heap implementation
for DFSSSP and SSSP. Instead, a 4-ary heap of the component
library will be used. This adds approx. 5-10% runtime overhead
for SSSP (which will still be in the range of MinHop routing).
For DFSSSP, the overhead to compute the routes with Dijkstra's
algorithm and 4-ary heap can be ignored, since most time is
spend to remove cyclic dependencies in the CDG after the routes
have been calculated.

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>