osm_ucast_ftree.c: Mark HCA ports connected to unranked switches
With this change, when an unranked switched is removed, HCA ports
connected to the switch are marked in order to not be routed by the
routing algorithm.

This change fixes a bug where there is a HCA which is connected to
at least two switches, and one of them is removed after the algorithm
did not decide on its rank. Since the HCA is also connected to a switch
with a rank, the HCA is not removed and the algorithm will try to route
all its ports.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>