dfsssp: Ignore throttled links in fabric
Throttled links, either because they are faulty or throttled intentionally,
aren't good for the achievable performance of an HPC system.
During the discovery process, (DF)SSSP analyzes the theoretical maximum
enabled link speed, which both ends of a peer link support and compares it
to the actual link speed. If these numbers don't match, then the link is
ignored as potential path for routes between nodes. A similar comparison
is done for link width.

The implementation of the link state analyzer is also available to other
routing engines and the patch adds an option in OpenSM conf to control
the feature (default is 'disabled' to preserve current behavior).

Signed-off-by: Jens Domke <domke.j.aa@m.titech.ac.jp>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>