Add timeout parameter for SM class set transactions
Currently, this timeout is 0 (default) for all SM sets.
It was tested by setting the set LFT timeout to 5 seconds
and breaking the DR path so it would timeout.

Subsequent patch(es) will add in which set transactions
will use non default timeouts. Optimized SL2VLMappingTable
sets are one candidate for this but there are others.
It may also be extended to SM class get transactions.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>