Add option and support for only using the original extended SA rates
prior to the addition of 2x link width and HDR link speed.
SA rate is a field in SA PathRecord, MultiPathRecord, and MCMemberRecord.

The original values are up through and including 18: 300 Gbps (12x EDR).
This option is needed for subnets with old kernels/drivers
that don't understand the new rates added for 2x link
width and/or HDR link speed support (19-22).

Signed-off-by: Hal Rosenstock <hal@mellanox.com>