osm_[multi]path_record.c: Fix a couple of edge cases with new 2x/HDR SA rates
Ports which do not support 2x/HDR cannot handle 28 Gbps (18 - EDR 2x)
or 50 Gbps (19 - FDR 2x) rates so convert these rates to next lower
acceptable rate. In case of 28 Gbps, it is 25 Gbps (15) for EDR  1x
when extended speeds is supported or 20 Gbps (6) when only the original
speeds are supported. In case of 50 Gbps, it is always 40 Gbps (7).

Signed-off-by: Hal Rosenstock <hal@mellanox.com>