osm_db_files.c: Remove an obsolete GUID-length check on osm_db_restore
The code currently checks that the first word (the key) is up to
MAX_GUID_LENGTH which is defined as 32. There is no real reason
to do that check, and apparently the only purpose was to make sure
the key is a valid GUID. Since for the last few decades,
the key is not necessarily a GUID but can also be a mixture of
elements (for example GUID:port) this check seems pointless.

Signed-off-by: Elad Weiss <eladw@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>