osm_link_mgr.c: Fix checking if port support link width 2x
Bug description:
When force_link_width option is enabled, OpenSM checks for switch
external ports if the port support link width 2x in PortInfo of the
external port, instead of PortInfo of switch port 0.

With this change, when OpenSM checks if a switch supports link width 2x,
it checks in PortInfo of port 0 of the device.

Signed-off-by: Daniel Klein <danielk@mellanox.com>